adding separated test runners
