+++ tools/please_pex/BUILD
@@ -3,7 +3,9 @@ filegroup(
     srcs = [
         'pex_main.py',
         'pex_run.py',
+        'pytest.py',
         'test_main.py',
+        'unittest.py',
     ],
     visibility = ['//tools/please_pex/pex:all'],
 )
