+++ tools/please_pex/pytest.py
@@ -1,0 +1,6 @@
+import pytest
+
+
+def run_tests(test_names):
+    """Runs tests using pytest, returns the number of failures."""
+    return 1
